functions/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.7.0
blobfile==2.0.1
bs4==0.0.1
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
cloudevents==1.9.0
comm==0.1.2
contourpy==1.0.7
cryptography==41.0.5
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
deprecation==2.1.0
distro==1.8.0
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
filelock==3.9.0
firebase-admin==6.2.0
firebase-functions==0.1.2
Flask==2.3.3
Flask-Cors==4.0.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.10.0
functions-framework==3.4.0
google-api-core==2.14.0
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-events==0.11.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.1
huggingface-hub==0.19.1
idna==3.4
ipykernel==6.20.1
ipython==8.10.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==5.1.3
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.6.3
matplotlib-inline==0.1.6
msgpack==1.0.7
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.1
openai==0.26.1
packaging==23.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.4.12
platformdirs==2.6.2
plotly==5.12.0
prompt-toolkit==3.0.36
proto-plus==1.22.3
protobuf==4.25.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.17
Pygments==2.15.0
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.31.0
rsa==4.9
safetensors==0.4.0
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tenacity==8.1.0
threadpoolctl==3.1.0
tiktoken==0.1.2
tokenizers==0.13.2
tornado==6.3.3
tqdm==4.64.1
traitlets==5.8.1
transformers==4.30.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
watchdog==3.0.0
wcwidth==0.2.5
Werkzeug==3.0.1
yarg==0.1.9
yarl==1.8.2